+++ boa_restrictor/__init__.py
@@ -1,3 +1,3 @@
 """A custom Python and Django linter from Ambient"""
 
-__version__ = "1.13.3"
+__version__ = "1.14.1"